wanita/requirements.txt
gunicorn
Flask
google-cloud-texttospeech
ffmpeg-python
